refactor: according to test senarios & import issue
