+++ src/techlog_article/article/application/port/in_/_read.py
@@ -5,7 +5,7 @@ from datetime import datetime
 
 from pydantic import BaseModel, Field, validator
 
-from domain.article import Article
+from ....domain import Article
 
 
 class ReadArticleResponse(BaseModel):
