+++ src/techlog_article/article/domain/__init__.py
@@ -1,0 +1,3 @@
+from ._article import Article
+
+__all__ = ["Article"]
