+++ src/techlog_article/article/domain/_article.py
