+++ src/techlog_article/auth/domain/__init__.py
@@ -1,0 +1,3 @@
+from ._auth import Auth
+
+__all__ = ["Auth"]
