+++ src/techlog_article/user/__init__.py
@@ -1,3 +1,3 @@
-from .domain.user import User
+from .domain import User
 
 __all__ = ["User"]
