+++ src/techlog_article/user/domain/__init__.py
@@ -1,0 +1,3 @@
+from ._user import User
+
+__all__ = ["User"]
