--- tests/test_example.py
@@ -1,2 +1,0 @@
-def test_example():
-    assert 1 + 1 == 2
